collect fuzz tests
