+++ allandsorted_test.go
@@ -11,106 +11,6 @@ import (
 	"testing"
 )
 
-// FuzzAllSorted tests that AllSorted() visits all prefixes in natural CIDR sort order
-func FuzzAllSorted(f *testing.F) {
-	// Seed with some initial test cases
-	f.Add(uint64(12345), 100)
-	f.Add(uint64(67890), 250)
-	f.Add(uint64(11111), 500)
-
-	f.Fuzz(func(t *testing.T, seed uint64, count int) {
-		// Bound the test size to reasonable limits
-		if count < 10 || count > 1000 {
-			t.Skip("count out of range")
-		}
-
-		// Generate random prefixes using the existing utility
-		prng := rand.New(rand.NewPCG(seed, 42))
-		prefixItems := randomPrefixes(prng, count)
-
-		type tabler interface {
-			Insert(netip.Prefix, int)
-			AllSorted() iter.Seq2[netip.Prefix, int]
-		}
-
-		// Test all table types that support AllSorted
-		tables := []struct {
-			name    string
-			builder func() tabler
-		}{
-			{"Fast", func() tabler { return new(Fast[int]) }},
-			{"Table", func() tabler { return new(Table[int]) }},
-			{"liteTable", func() tabler { return new(liteTable[int]) }},
-		}
-
-		for _, tt := range tables {
-			t.Run(tt.name, func(t *testing.T) {
-				t.Parallel()
-
-				tbl := tt.builder()
-
-				// Insert all prefixes with their values
-				for _, item := range prefixItems {
-					tbl.Insert(item.pfx, item.val)
-				}
-
-				// Collect all prefixes from AllSorted
-				sortedPrefixes := make([]netip.Prefix, 0, count)
-				sortedValues := make([]int, 0, count)
-
-				for pfx, val := range tbl.AllSorted() {
-					sortedPrefixes = append(sortedPrefixes, pfx)
-					sortedValues = append(sortedValues, val)
-				}
-
-				// Verify we got exactly the same number of prefixes
-				if len(sortedPrefixes) != len(prefixItems) {
-					t.Fatalf("%s: Expected %d prefixes, got %d", tt.name, len(prefixItems), len(sortedPrefixes))
-				}
-
-				// Verify all prefixes are in natural CIDR sort order using existing cmpPrefix function
-				for i := 1; i < len(sortedPrefixes); i++ {
-					if cmpPrefix(sortedPrefixes[i-1], sortedPrefixes[i]) > 0 {
-						t.Fatalf("%s: CIDR sort order violated at index %d: %v should come before %v",
-							tt.name, i-1, sortedPrefixes[i-1], sortedPrefixes[i])
-					}
-				}
-
-				// Verify we can find each original prefix in the sorted results
-				// Create a map for O(1) lookup verification
-				resultMap := make(map[netip.Prefix]int, count)
-				for i, pfx := range sortedPrefixes {
-					resultMap[pfx] = sortedValues[i]
-				}
-
-				for _, originalItem := range prefixItems {
-					val, found := resultMap[originalItem.pfx]
-					if !found {
-						t.Fatalf("%s: Original prefix %v not found in AllSorted results", tt.name, originalItem.pfx)
-					}
-
-					// liteTable has no payload
-					if _, ok := tbl.(*liteTable[int]); !ok {
-						if val != originalItem.val {
-							t.Fatalf("%s: Original prefix %v has wrong value: expected %d, got %d",
-								tt.name, originalItem.pfx, originalItem.val, val)
-						}
-					}
-				}
-
-				// Verify no duplicates in results (since randomPrefixes guarantees no duplicates)
-				seen := make(map[netip.Prefix]bool, count)
-				for _, pfx := range sortedPrefixes {
-					if seen[pfx] {
-						t.Fatalf("%s: Duplicate prefix %v found in AllSorted results", tt.name, pfx)
-					}
-					seen[pfx] = true
-				}
-			})
-		}
-	})
-}
-
 // TestAll tests All with random samples
 func TestAll(t *testing.T) {
 	t.Parallel()
