+++ iterators_test.go
@@ -325,297 +325,3 @@ func TestAllSorted4AllSorted6_Lite(t *testing.T) {
 	verifySortedCIDR(t, s4)
 	verifySortedCIDR(t, s6)
 }
-
-// ---------------- Fuzz: Subnets ----------------
-
-//nolint:gocyclo
-func FuzzSubnets(f *testing.F) {
-	// Seed corpus
-	f.Add(uint64(12345), 150, 30)
-	f.Add(uint64(67890), 400, 60)
-	f.Add(uint64(54321), 800, 100)
-	// Edge-case leaning seeds
-	f.Add(uint64(0), 64, 16)    // bias towards small sets
-	f.Add(^uint64(0), 1024, 64) // large sets
-
-	f.Fuzz(func(t *testing.T, seed uint64, n, nq int) {
-		if n < 10 || n > 5000 || nq < 1 || nq > 200 {
-			t.Skip("bounds")
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 13))
-		pfxs := randomPrefixes(prng, n)
-		queries := randomPrefixes(prng, nq)
-
-		// Table
-		{
-			tbl := new(Table[int])
-			for i, it := range pfxs {
-				tbl.Insert(it.pfx, i)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSubnetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range tbl.Subnets(q.pfx) {
-					if got[p] {
-						t.Fatalf("Table.Subnets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("Table.Subnets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("Table.Subnets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-
-		// Fast
-		{
-			ft := new(Fast[int])
-			for i, it := range pfxs {
-				ft.Insert(it.pfx, i)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSubnetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range ft.Subnets(q.pfx) {
-					if got[p] {
-						t.Fatalf("Fast.Subnets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("Fast.Subnets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("Fast.Subnets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-
-		// liteTable
-		{
-			lt := new(liteTable[int])
-			for i, it := range pfxs {
-				lt.Insert(it.pfx, i)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSubnetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range lt.Subnets(q.pfx) {
-					if got[p] {
-						t.Fatalf("liteTable.Subnets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("liteTable.Subnets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("liteTable.Subnets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-
-		// Lite
-		{
-			l := new(Lite)
-			for _, it := range pfxs {
-				l.Insert(it.pfx)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSubnetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range l.Subnets(q.pfx) {
-					if got[p] {
-						t.Fatalf("Lite.Subnets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("Lite.Subnets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("Lite.Subnets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-	})
-}
-
-// ---------------- Fuzz: Supernets ----------------
-
-//nolint:gocyclo
-func FuzzSupernets(f *testing.F) {
-	// Seed corpus
-	f.Add(uint64(222), 150, 30)
-	f.Add(uint64(333), 400, 60)
-	f.Add(uint64(444), 800, 100)
-	// Edge-case leaning seeds
-	f.Add(uint64(0), 64, 16)    // bias towards small sets
-	f.Add(^uint64(0), 1024, 64) // large sets
-
-	f.Fuzz(func(t *testing.T, seed uint64, n, nq int) {
-		if n < 10 || n > 5000 || nq < 1 || nq > 200 {
-			t.Skip("bounds")
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 17))
-		pfxs := randomPrefixes(prng, n)
-		queries := randomPrefixes(prng, nq)
-
-		// Table
-		{
-			tbl := new(Table[int])
-			for i, it := range pfxs {
-				tbl.Insert(it.pfx, i)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSupernetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range tbl.Supernets(q.pfx) {
-					if got[p] {
-						t.Fatalf("Table.Supernets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("Table.Supernets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("Table.Supernets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-
-		// Fast
-		{
-			ft := new(Fast[int])
-			for i, it := range pfxs {
-				ft.Insert(it.pfx, i)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSupernetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range ft.Supernets(q.pfx) {
-					if got[p] {
-						t.Fatalf("Fast.Supernets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("Fast.Supernets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("Fast.Supernets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-
-		// liteTable
-		{
-			lt := new(liteTable[int])
-			for i, it := range pfxs {
-				lt.Insert(it.pfx, i)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSupernetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range lt.Supernets(q.pfx) {
-					if got[p] {
-						t.Fatalf("liteTable.Supernets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("liteTable.Supernets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("liteTable.Supernets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-
-		// Lite
-		{
-			l := new(Lite)
-			for _, it := range pfxs {
-				l.Insert(it.pfx)
-			}
-			for _, q := range queries {
-				want := map[netip.Prefix]bool{}
-				for _, it := range pfxs {
-					if isSupernetOf(it.pfx, q.pfx) {
-						want[it.pfx] = true
-					}
-				}
-				got := map[netip.Prefix]bool{}
-				for p := range l.Supernets(q.pfx) {
-					if got[p] {
-						t.Fatalf("Lite.Supernets duplicate: %v", p)
-					}
-					got[p] = true
-				}
-				if len(got) != len(want) {
-					t.Fatalf("Lite.Supernets size mismatch for %v: want %d got %d", q.pfx, len(want), len(got))
-				}
-				for p := range want {
-					if !got[p] {
-						t.Fatalf("Lite.Supernets missing %v for %v", p, q.pfx)
-					}
-				}
-			}
-		}
-	})
-}
