+++ overlaps_test.go
@@ -31,179 +31,6 @@ func expectedOverlapsPrefix(as []netip.Prefix, q netip.Prefix) bool {
 	return false
 }
 
-// FuzzOverlaps validates Overlaps(o) vs ground-truth for all table types.
-func FuzzOverlaps(f *testing.F) {
-	// Seed corpus
-	f.Add(uint64(12345), 50, 50)
-	f.Add(uint64(67890), 150, 75)
-	f.Add(uint64(22222), 300, 300)
-
-	f.Fuzz(func(t *testing.T, seed uint64, n1, n2 int) {
-		if n1 < 1 || n1 > 1000 || n2 < 1 || n2 > 1000 {
-			t.Skip("counts out of range")
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 42))
-		aItems := randomPrefixes(prng, n1) // invariant: no duplicates
-		bItems := randomPrefixes(prng, n2)
-
-		aPfx := make([]netip.Prefix, len(aItems))
-		for i, it := range aItems {
-			aPfx[i] = it.pfx
-		}
-		bPfx := make([]netip.Prefix, len(bItems))
-		for i, it := range bItems {
-			bPfx[i] = it.pfx
-		}
-
-		gt := expectedOverlaps(aPfx, bPfx)
-
-		// Table
-		{
-			t1 := new(Table[int])
-			t2 := new(Table[int])
-			for _, it := range aItems {
-				t1.Insert(it.pfx, it.val)
-			}
-			for _, it := range bItems {
-				t2.Insert(it.pfx, it.val)
-			}
-			if got := t1.Overlaps(t2); got != gt {
-				t.Fatalf("Table.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-
-		// Fast
-		{
-			f1 := new(Fast[int])
-			f2 := new(Fast[int])
-			for _, it := range aItems {
-				f1.Insert(it.pfx, it.val)
-			}
-			for _, it := range bItems {
-				f2.Insert(it.pfx, it.val)
-			}
-			if got := f1.Overlaps(f2); got != gt {
-				t.Fatalf("Fast.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-
-		// liteTable (payload-less semantics—overlap depends only on prefixes)
-		{
-			lt1 := new(liteTable[int])
-			lt2 := new(liteTable[int])
-			for _, it := range aItems {
-				lt1.Insert(it.pfx, 0)
-			}
-			for _, it := range bItems {
-				lt2.Insert(it.pfx, 0)
-			}
-			if got := lt1.Overlaps(lt2); got != gt {
-				t.Fatalf("liteTable.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-
-		// Lite (prefix-only)
-		{
-			L1 := new(Lite)
-			L2 := new(Lite)
-			for _, it := range aItems {
-				L1.Insert(it.pfx)
-			}
-			for _, it := range bItems {
-				L2.Insert(it.pfx)
-			}
-			if got := L1.Overlaps(L2); got != gt {
-				t.Fatalf("Lite.Overlaps mismatch: want %v, got %v", gt, got)
-			}
-		}
-	})
-}
-
-// FuzzOverlapsPrefix validates OverlapsPrefix(q) vs ground-truth for all table types.
-// For each query, it checks that OverlapsPrefix matches any(a.Overlaps(q)).
-func FuzzOverlapsPrefix(f *testing.F) {
-	// Seed corpus
-	f.Add(uint64(111), 200, 60)
-	f.Add(uint64(222), 600, 120)
-	f.Add(uint64(333), 1200, 180)
-
-	f.Fuzz(func(t *testing.T, seed uint64, n, nq int) {
-		if n < 1 || n > 4000 || nq < 1 || nq > 400 {
-			t.Skip("counts out of range")
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 7))
-		items := randomPrefixes(prng, n)  // table contents
-		query := randomPrefixes(prng, nq) // queries, use .pfx
-
-		contents := make([]netip.Prefix, len(items))
-		for i, it := range items {
-			contents[i] = it.pfx
-		}
-
-		// Table
-		{
-			tbl := new(Table[int])
-			for _, it := range items {
-				tbl.Insert(it.pfx, it.val)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := tbl.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("Table.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-
-		// Fast
-		{
-			fst := new(Fast[int])
-			for _, it := range items {
-				fst.Insert(it.pfx, it.val)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := fst.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("Fast.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-
-		// liteTable (presence only)
-		{
-			lt := new(liteTable[int])
-			for _, it := range items {
-				lt.Insert(it.pfx, 0)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := lt.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("liteTable.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-
-		// Lite (prefix-only)
-		{
-			L := new(Lite)
-			for _, it := range items {
-				L.Insert(it.pfx)
-			}
-			for _, q := range query {
-				gt := expectedOverlapsPrefix(contents, q.pfx)
-				got := L.OverlapsPrefix(q.pfx)
-				if got != gt {
-					t.Fatalf("Lite.OverlapsPrefix(%v) mismatch: want %v, got %v", q.pfx, gt, got)
-				}
-			}
-		}
-	})
-}
-
 // TestOverlapsDeterministic_EdgeCases exercises thorough edge coverage for Overlaps across all table types:
 // - IPv6: siblings (/33 under /32), default route (::/0), host routes (/128), /127 semantics
 // - IPv4: default route (0.0.0.0/0), host routes (/32), /31 semantics, adjacent siblings
