+++ tablemodify_test.go
@@ -1,7 +1,6 @@
 package bart
 
 import (
-	"math/rand/v2"
 	"net/netip"
 	"testing"
 )
@@ -831,261 +830,4 @@ func TestLiteTableVsTableComparison(t *testing.T) {
 	if regularTable.Size() != 1 || lite.Size() != 1 {
 		t.Error("Both tables should have size 1")
 	}
-}
-
-// Comprehensive fuzz test for Lite Modify operations
-func FuzzLiteModifyComprehensive(f *testing.F) {
-	seeds := []struct {
-		seed  uint64
-		count int
-		op    uint8 // 0=insert, 2=delete, 3=no-op (skip update since no payload)
-	}{
-		{12345, 50, 0},
-		{67890, 25, 2},
-		{11111, 75, 3},
-		{22222, 30, 0},
-		{33333, 10, 2},
-	}
-
-	for _, seed := range seeds {
-		f.Add(seed.seed, seed.count, seed.op)
-	}
-
-	f.Fuzz(func(t *testing.T, seed uint64, count int, op uint8) {
-		if count < 5 || count > 100 {
-			return
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 42))
-		prefixItems := randomPrefixes(prng, count)
-
-		if len(prefixItems) == 0 {
-			return
-		}
-
-		targetIdx := prng.IntN(len(prefixItems))
-		targetPrefix := prefixItems[targetIdx].pfx
-
-		lite := &Lite{}
-
-		// Setup: Insert first half of prefixes
-		halfCount := len(prefixItems) / 2
-		for i := range halfCount {
-			item := prefixItems[i]
-			lite.Modify(item.pfx, func(bool) bool {
-				return false // insert
-			})
-		}
-
-		initialSize := lite.Size()
-		initialFound := lite.Get(targetPrefix)
-
-		// Expected outcome tracking
-		var expectedSize int
-		var expectedFound bool
-
-		// Execute modify operation - skip update ops since Lite has no meaningful payload
-		lite.Modify(targetPrefix, func(found bool) bool {
-			// Verify callback parameters
-			if found != initialFound {
-				t.Errorf("callback found=%v, but actual found=%v", found, initialFound)
-			}
-
-			// Map op to valid operations for liteTable (no update)
-			switch op % 3 { // Use mod 3 to skip update operation
-			case 0: // insert if not found
-				if !found {
-					expectedSize = initialSize + 1
-					expectedFound = true
-					return false // insert
-				}
-				// Already exists, no change
-				expectedSize = initialSize
-				expectedFound = true
-				return false // keep existing
-
-			case 1: // delete if found (mod 3 case 1)
-				if found {
-					expectedSize = initialSize - 1
-					expectedFound = false
-					return true // delete
-				}
-				// Not found, no-op
-				expectedSize = initialSize
-				expectedFound = false
-				return true // no-op with del=true
-
-			case 2: // no-op always (mod 3 case 2)
-				expectedSize = initialSize
-				expectedFound = found
-
-				if found {
-					return false // keep existing
-				} else {
-					return true // no-op with del=true
-				}
-			}
-
-			panic("unreachable")
-		})
-
-		// Verify results
-		if lite.Size() != expectedSize {
-			t.Errorf("Size inconsistent: got %d, expected %d (op=%d, initialFound=%v)",
-				lite.Size(), expectedSize, op%3, initialFound)
-		}
-
-		actualFound := lite.Get(targetPrefix)
-		if actualFound != expectedFound {
-			t.Errorf("Get found inconsistent: got %v, expected %v (op=%d, initialFound=%v)",
-				actualFound, expectedFound, op%3, initialFound)
-		}
-	})
-}
-
-// Comprehensive fuzz test using existing random prefix generators
-func FuzzTableModifyComprehensive(f *testing.F) {
-	seeds := []struct {
-		seed  uint64
-		count int
-		value int
-		op    uint8
-	}{
-		{12345, 50, 100, 0},
-		{67890, 25, 200, 1},
-		{11111, 75, 300, 2},
-		{22222, 30, 400, 3},
-		{33333, 10, 500, 0},
-	}
-
-	for _, seed := range seeds {
-		f.Add(seed.seed, seed.count, seed.value, seed.op)
-	}
-
-	f.Fuzz(func(t *testing.T, seed uint64, count int, value int, op uint8) {
-		if count < 5 || count > 100 {
-			return
-		}
-
-		prng := rand.New(rand.NewPCG(seed, 42))
-		prefixItems := randomPrefixes(prng, count)
-
-		if len(prefixItems) == 0 {
-			return
-		}
-
-		targetIdx := prng.IntN(len(prefixItems))
-		targetPrefix := prefixItems[targetIdx].pfx
-
-		tables := []interface {
-			Modify(netip.Prefix, func(int, bool) (int, bool))
-			Get(netip.Prefix) (int, bool)
-			Size() int
-		}{
-			&Table[int]{},
-			&Fast[int]{},
-		}
-
-		for tableIdx, table := range tables {
-			// Setup: Insert first half of prefixes
-			halfCount := len(prefixItems) / 2
-			for i := range halfCount {
-				item := prefixItems[i]
-				table.Modify(item.pfx, func(_ int, _ bool) (int, bool) {
-					return item.val, false
-				})
-			}
-
-			initialSize := table.Size()
-			initialVal, initialFound := table.Get(targetPrefix)
-
-			var expectedSize int
-			var expectedFound bool
-			var expectedVal int
-
-			table.Modify(targetPrefix, func(val int, found bool) (int, bool) {
-				// Verify callback parameters match actual state
-				if found != initialFound {
-					t.Errorf("Table %d: callback found=%v, but actual found=%v",
-						tableIdx, found, initialFound)
-				}
-				if found && val != initialVal {
-					t.Errorf("Table %d: callback val=%v, but actual val=%v",
-						tableIdx, val, initialVal)
-				}
-
-				switch op % 4 {
-				case 0: // insert if not found
-					if !found {
-						expectedSize = initialSize + 1
-						expectedFound = true
-						expectedVal = value
-						return value, false // insert new value
-					}
-					// Already exists, keep existing
-					expectedSize = initialSize
-					expectedFound = true
-					expectedVal = val
-					return val, false // no change
-
-				case 1: // update if found
-					if found {
-						expectedSize = initialSize
-						expectedFound = true
-						expectedVal = value
-						return value, false // update to new value
-					}
-					// Not found, no-op
-					expectedSize = initialSize
-					expectedFound = false
-					expectedVal = 0
-					return 0, true // del=true means no-op
-
-				case 2: // delete if found
-					if found {
-						expectedSize = initialSize - 1
-						expectedFound = false
-						expectedVal = 0
-						return val, true // delete existing
-					}
-					// Not found, no-op
-					expectedSize = initialSize
-					expectedFound = false
-					expectedVal = 0
-					return 0, true // del=true means no-op
-
-				case 3: // no-op always
-					expectedSize = initialSize
-					expectedFound = found
-
-					if found {
-						expectedVal = val
-						return val, false // keep existing value unchanged
-					} else {
-						expectedVal = 0
-						return 0, true // del=true means no-op for non-existent
-					}
-				}
-
-				return 0, false
-			})
-
-			// Verify all results
-			if table.Size() != expectedSize {
-				t.Errorf("Table %d: Size inconsistent: got %d, expected %d (op=%d, initialFound=%v)",
-					tableIdx, table.Size(), expectedSize, op%4, initialFound)
-			}
-
-			actualVal, actualFound := table.Get(targetPrefix)
-			if actualFound != expectedFound {
-				t.Errorf("Table %d: Get found inconsistent: got %v, expected %v (op=%d, initialFound=%v)",
-					tableIdx, actualFound, expectedFound, op%4, initialFound)
-			}
-
-			if expectedFound && actualVal != expectedVal {
-				t.Errorf("Table %d: Get value inconsistent: got %v, expected %v (op=%d)",
-					tableIdx, actualVal, expectedVal, op%4)
-			}
-		}
-	})
 }
